feat: useLogout 적용
