+++ packages/client/src/hooks/useLogout.tsx
@@ -1,10 +1,14 @@
 import AuthAPI from '~/api/auth';
 
-const useLogout = async () => {
-  try {
-    await AuthAPI.logout();
-  } catch (error) {}
-  window.location.href = '/';
+const useLogout = () => {
+  const logout = async () => {
+    try {
+      await AuthAPI.logout();
+    } catch (error) {}
+    window.location.href = '/';
+  };
+
+  return logout;
 };
 
 export default useLogout;
